move `must_use` lint to a separate file
